team/requirements.txt
asgiref==3.7.2
backoff==2.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.61.1
django-captcha==1.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
idna==3.6
oauthlib==3.2.2
opencage==2.4.0
phonenumbers==8.13.28
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
